+++ crates/canic-core/src/domain/policy/pure/env.rs
@@ -1,6 +1,7 @@
 use crate::{
     domain::value::Principal,
     ids::{BuildNetwork, CanisterRole, SubnetRole},
+    model::env::ValidatedEnv,
 };
 use thiserror::Error as ThisError;
 
@@ -16,20 +17,6 @@ pub struct EnvInput {
     pub root_pid: Option<Principal>,
     pub canister_role: Option<CanisterRole>,
     pub parent_pid: Option<Principal>,
-}
-
-///
-/// ValidatedEnv
-///
-
-#[derive(Clone, Debug)]
-pub struct ValidatedEnv {
-    pub prime_root_pid: Principal,
-    pub subnet_role: SubnetRole,
-    pub subnet_pid: Principal,
-    pub root_pid: Principal,
-    pub canister_role: CanisterRole,
-    pub parent_pid: Principal,
 }
 
 ///